update main features
