--- src/main/__init__.py
@@ -1,3 +1,0 @@
-from .routes import main_bp  # noqa
-
-__all__ = ("main_bp",)
